announcements should scroll into view
